src/question/question.service.spec.ts
import { Answer } from "../../src/answer/entities/answer.entity";
import { Question } from "../../src/question/entities/question.entity";
// Import necessary modules and dependencies for your tests
import { Test, TestingModule } from '@nestjs/testing';
import { QuestionService } from '../../src/question/question.service';
import { getRepositoryToken } from '@nestjs/typeorm';
import { CreateQuestionInput } from "../../src/question/dto/create-question.input";
import { CreateAnswerInput } from "../../src/answer/dto/create-answer.input";

class MockRepository<T> {
    private entities: T[] = [];
  
    create(entity: T): T {
      this.entities.push(entity);
      return entity;
    }
  
    async save(entity: T): Promise<T> {
      return entity; // Simulate saving to the database
    }
  
    findOne(query: any): T | undefined {
      // Simulate finding an entity in the database
      return this.entities.find((entity) => {
        // Implement your custom logic here based on the query
        return /* Check if the entity matches the query criteria */;
      });
    }
  
    // Implement other methods you use in your service
  }
  
  // Create instances of the mock repository for Question and Answer entities
  const questionRepositoryMock = new MockRepository<Question>();
  const answerRepositoryMock = new MockRepository<Answer>();
  
  // Configure mock functions for the methods used in your service
  jest.spyOn(questionRepositoryMock, 'create').mockImplementation(questionRepositoryMock.create);
  jest.spyOn(questionRepositoryMock, 'save').mockImplementation(questionRepositoryMock.save);
  jest.spyOn(questionRepositoryMock, 'findOne').mockImplementation(questionRepositoryMock.findOne);
  jest.spyOn(answerRepositoryMock, 'create').mockImplementation(answerRepositoryMock.create);
  jest.spyOn(answerRepositoryMock, 'save').mockImplementation(answerRepositoryMock.save);
  jest.spyOn(answerRepositoryMock, 'findOne').mockImplementation(answerRepositoryMock.findOne);



// Define the MockRepository class and mock functions as shown in the previous answer

// Describe and write your test suite
describe('QuestionService', () => {
  let service: QuestionService;

  beforeEach(async () => {
    const module: TestingModule = await Test.createTestingModule({
      providers: [
        QuestionService,
        {
          provide: getRepositoryToken(Question),
          useValue: questionRepositoryMock,
        },
        {
          provide: getRepositoryToken(Answer),
          useValue: answerRepositoryMock,
        },
      ],
    }).compile();

    service = module.get<QuestionService>(QuestionService);
  });

  it('should create a question with answers', async () => {
    // Arrange
    const createQuestionInput = new CreateQuestionInput();
    createQuestionInput.question = 'What is the capital of France?';
    createQuestionInput.singleAnswer = true
    
    const answersInput = [
      new CreateAnswerInput(),
      // Add more answers as needed
    ];
    answersInput[0].answer = 'Paris';
    answersInput[0].correct = true;

    // Mock the behavior of the create method in the questionRepositoryMock
    questionRepositoryMock.create = jest.fn().mockReturnValue(createQuestionInput);

    // Mock the behavior of the save method in the questionRepositoryMock
    questionRepositoryMock.save = jest.fn().mockReturnValue(createQuestionInput);

    // Mock the behavior of the create method in the answerRepositoryMock
    answerRepositoryMock.create = jest.fn().mockImplementation((answer) => ({
      ...answer,
      id: 'generated-answer-id', // Simulate an ID generation
    }));

    // Mock the behavior of the save method in the answerRepositoryMock
    answerRepositoryMock.save = jest.fn().mockImplementation((answer) => answer);

    // Act
    const createdQuestion = await service.create(createQuestionInput, answersInput);

    // Assert
    expect(createdQuestion).toEqual(createQuestionInput);
    expect(createdQuestion.answers).toHaveLength(answersInput.length);

    // Verify that the create method in the questionRepositoryMock was called with the input
    expect(questionRepositoryMock.create).toHaveBeenCalledWith(createQuestionInput);

    // Verify that the save method in the questionRepositoryMock was called
    expect(questionRepositoryMock.save).toHaveBeenCalledWith(createQuestionInput);

    // Verify that the create and save methods in the answerRepositoryMock were called for each answer
    for (const answerInput of answersInput) {
      expect(answerRepositoryMock.create).toHaveBeenCalledWith(answerInput);
      expect(answerRepositoryMock.save).toHaveBeenCalledWith(expect.objectContaining(answerInput));
    }
  });

  // Write your test cases here, including the ones for create, update, checkAnswer, and remove methods.
});
